Lesson 12 from 11

+++ api/src/main/java/com/github/kastkest/spring_market/api/cart/CartDto.java
@@ -1,13 +1,14 @@
 package com.github.kastkest.spring_market.api.cart;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
 
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
-    public CartDto(List<CartItemDto> items, int totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
@@ -20,11 +21,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
